# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OmniSync",
    "version": "see VERSION"
  },
  "paths": {
...
    "/config": {
      "get": {
        "summary": "Get Config",
        "description": "Return the current global configuration.",
        "operationId": "get_config_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalConfigResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Config",
        "description": "Update global configuration with partial values.",
        "operationId": "update_config_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GlobalConfigUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalConfigResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/notifications/config": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "Get Notification Config",
        "description": "Read current notification channel preferences (credentials are never returned).",
        "operationId": "get_notification_config_notifications_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationConfigResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "notifications"
        ],
        "summary": "Update Notification Config",
        "description": "Update notification channel preferences and persist to config.toml.\n\nPartial: only the channels and fields given change; a field left out\nkeeps its current value.",
        "operationId": "update_notification_config_notifications_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationConfigUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationConfigResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/notifications/channels/status": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "Get Channel Status",
        "description": "Status of every registered channel: availability, what is missing, the detected host.",
        "operationId": "get_channel_status_notifications_channels_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelStatusResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/notifications/vapid-public-key": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "Get Vapid Public Key",
        "description": "Return the VAPID public key for browser push subscription.",
        "operationId": "get_vapid_public_key_notifications_vapid_public_key_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VapidPublicKeyResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/notifications/push-subscription": {
      "post": {
        "tags": [
          "notifications"
        ],
        "summary": "Subscribe Push",
        "description": "Register a browser push subscription (idempotent: a known endpoint is updated).",
        "operationId": "subscribe_push_notifications_push_subscription_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushSubscriptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Subscribe Push Notifications Push Subscription Post"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "notifications"
        ],
        "summary": "Unsubscribe Push",
        "description": "Remove a browser push subscription.",
        "operationId": "unsubscribe_push_notifications_push_subscription_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushUnsubscribeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Unsubscribe Push Notifications Push Subscription Delete"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/notifications/history": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "Get Notification History",
        "description": "Return paginated notification history.",
        "operationId": "get_notification_history_notifications_history_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "profile",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by profile slug",
              "title": "Profile"
            },
            "description": "Filter by profile slug"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationHistoryResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/notifications/test": {
      "post": {
        "tags": [
          "notifications"
        ],
        "summary": "Send Test Notification",
        "description": "Send a test notification through all enabled channels, or one given channel.\n\nA test of one channel is sent even if the channel is turned off, so it\ncan be checked before it is turned on.",
        "operationId": "send_test_notification_notifications_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/TestNotificationRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestNotificationResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "Server error (codes: docs/api-errors.md)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChannelStatusResponse": {
        "properties": {
          "channels": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ChannelStatusInfo"
            },
            "type": "object",
            "title": "Channels"
          },
          "config_error": {
            "type": "boolean",
            "title": "Config Error",
            "default": false
          },
          "unknown_channels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unknown Channels",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "channels"
        ],
        "title": "ChannelStatusResponse"
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "What went wrong, for people. Show this text."
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable machine-readable code (snake_case) to branch on."
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Extra data for some codes, e.g. errors, invalid_paths, names, replacement, retry_after."
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id",
            "description": "The request's id (also the X-Request-ID header); the server log lines of the request carry it."
          }
        },
        "type": "object",
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorResponse",
        "description": "The body of every error answer (4xx and 5xx); docs/api-errors.md lists the codes."
      },
...
      "GlobalConfigResponse": {
        "properties": {
          "log_level": {
            "type": "string",
            "title": "Log Level",
            "default": "INFO"
          },
          "history_days": {
            "type": "integer",
            "title": "History Days",
            "default": 90
          }
        },
        "type": "object",
        "title": "GlobalConfigResponse"
      },
      "GlobalConfigUpdateRequest": {
        "properties": {
          "log_level": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "DEBUG",
                  "INFO",
                  "WARNING",
                  "ERROR",
                  "CRITICAL"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Log Level"
          },
          "history_days": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 3650.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "History Days"
          }
        },
        "type": "object",
        "title": "GlobalConfigUpdateRequest"
      },
...
      "NotificationConfigResponse": {
        "properties": {
          "channels": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ChannelConfig"
            },
            "type": "object",
            "title": "Channels"
          }
        },
        "type": "object",
        "required": [
          "channels"
        ],
        "title": "NotificationConfigResponse"
      },
      "NotificationConfigUpdateRequest": {
        "properties": {
          "channels": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/ChannelConfigUpdate"
                },
                "propertyNames": {
                  "maxLength": 64
                },
                "type": "object",
                "maxProperties": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Channels"
          }
        },
        "type": "object",
        "title": "NotificationConfigUpdateRequest"
      },
...
      "PushSubscriptionRequest": {
        "properties": {
          "endpoint": {
            "type": "string",
            "maxLength": 2048,
            "title": "Endpoint"
          },
          "keys": {
            "additionalProperties": {
              "type": "string",
              "maxLength": 512
            },
            "propertyNames": {
              "maxLength": 32
            },
            "type": "object",
            "maxProperties": 8,
            "title": "Keys"
          }
        },
        "type": "object",
        "required": [
          "endpoint",
          "keys"
        ],
        "title": "PushSubscriptionRequest"
      },
...
      "VapidPublicKeyResponse": {
        "properties": {
          "public_key": {
            "type": "string",
            "title": "Public Key"
          }
        },
        "type": "object",
        "required": [
          "public_key"
        ],
        "title": "VapidPublicKeyResponse"
      },
...
    }
  }
}